repositories and middleware commit
fill all repositories,
add column admin in table users,
add Admin middleware - only admin can change DB

+++ src/routes/web.php
@@ -10,6 +10,7 @@ use App\Http\Controllers\StartController;
 use App\Http\Controllers\HeadingController;
 use App\Http\Controllers\IssueController;
 use App\Http\Controllers\VideoController;
+use App\Http\Middleware\Admin;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Route;
 use \App\Http\Controllers\StateController;
@@ -44,14 +45,14 @@ Route::group([
     'prefix' => 'issues'
 ], function () {
     Route::get('/', [IssueController::class, 'index'])->name('index');
-    Route::get('/create', [IssueController::class, 'create'])->name('create');
-    Route::post('/', [IssueController::class, 'store'])->name('store');
+    Route::get('/create', [IssueController::class, 'create'])->name('create')->middleware('admin');
+    Route::post('/', [IssueController::class, 'store'])->name('store')->middleware('admin');
     Route::post('/search', [IssueController::class, 'search'])->name('search');
     Route::get('/{issue}/', [IssueController::class, 'show'])->name('show');
-    Route::get('/{issue}/edit', [IssueController::class, 'edit'])->name('edit');
-    Route::put('/{issue}', [IssueController::class, 'update'])->name('edit');
-    Route::delete('/{id}', [IssueController::class, 'destroy'])->name('destroy');
-    Route::put('/recover/{id}', [IssueController::class, 'recover'])->name('recover');
+    Route::get('/{issue}/edit', [IssueController::class, 'edit'])->name('edit')->middleware('admin');
+    Route::put('/{issue}', [IssueController::class, 'update'])->name('update')->middleware('admin');
+    Route::delete('/{id}', [IssueController::class, 'destroy'])->name('destroy')->middleware('admin');;
+    Route::put('/recover/{id}', [IssueController::class, 'recover'])->name('recover')->middleware('admin');;
 });
 
 Route::group([
@@ -59,13 +60,13 @@ Route::group([
     'prefix' => 'categories'
 ], function () {
     Route::get('/', [CategoryController::class, 'index'])->name('index');
-    Route::get('/create', [CategoryController::class, 'create'])->name('create');
-    Route::post('/', [CategoryController::class, 'store'])->name('store');
+    Route::get('/create', [CategoryController::class, 'create'])->name('create')->middleware('admin');
+    Route::post('/', [CategoryController::class, 'store'])->name('store')->middleware('admin');
     Route::post('/search', [CategoryController::class, 'search'])->name('search');
     Route::get('/{category}/', [CategoryController::class, 'show'])->name('show');
-    Route::get('/{category}/edit', [CategoryController::class, 'edit'])->name('edit');
-    Route::put('/{category}', [CategoryController::class, 'update'])->name('edit');
-    Route::delete('/{category}', [CategoryController::class, 'destroy'])->name('destroy');
+    Route::get('/{category}/edit', [CategoryController::class, 'edit'])->name('edit')->middleware('admin');
+    Route::put('/{category}', [CategoryController::class, 'update'])->name('update')->middleware('admin');
+    Route::delete('/{category}', [CategoryController::class, 'destroy'])->name('destroy')->middleware('admin');
 });
 
 Route::group([
@@ -73,12 +74,12 @@ Route::group([
     'prefix' => 'media'
 ], function () {
     Route::get('/', [MediaFolderController::class, 'index'])->name('index');
-    Route::get('/create', [MediaFolderController::class, 'create'])->name('create');
-    Route::post('/', [MediaFolderController::class, 'store'])->name('store');
+    Route::get('/create', [MediaFolderController::class, 'create'])->name('create')->middleware('admin');
+    Route::post('/', [MediaFolderController::class, 'store'])->name('store')->middleware('admin');
     Route::get('/{media}/', [MediaFolderController::class, 'show'])->name('show');
-    Route::get('/{media}/edit', [MediaFolderController::class, 'edit'])->name('edit');
-    Route::put('/{media}', [MediaFolderController::class, 'update'])->name('edit');
-    Route::delete('/{media}', [MediaFolderController::class, 'destroy'])->name('destroy');
+    Route::get('/{media}/edit', [MediaFolderController::class, 'edit'])->name('edit')->middleware('admin');
+    Route::put('/{media}', [MediaFolderController::class, 'update'])->name('update')->middleware('admin');
+    Route::delete('/{media}', [MediaFolderController::class, 'destroy'])->name('destroy')->middleware('admin');
 });
 
 Route::group([
@@ -86,12 +87,12 @@ Route::group([
     'prefix' => 'photos'
 ], function () {
     Route::get('/', [PhotoController::class, 'index'])->name('index');
-    Route::get('/create', [PhotoController::class, 'create'])->name('create');
-    Route::post('/', [PhotoController::class, 'store'])->name('store');
+    Route::get('/create', [PhotoController::class, 'create'])->name('create')->middleware('admin');
+    Route::post('/', [PhotoController::class, 'store'])->name('store')->middleware('admin');
     Route::get('/{media}/', [PhotoController::class, 'show'])->name('show');
-    Route::get('/{media}/edit', [PhotoController::class, 'edit'])->name('edit');
-    Route::put('/{media}', [PhotoController::class, 'update'])->name('edit');
-    Route::delete('/{media}', [PhotoController::class, 'destroy'])->name('destroy');
+    Route::get('/{media}/edit', [PhotoController::class, 'edit'])->name('edit')->middleware('admin');
+    Route::put('/{media}', [PhotoController::class, 'update'])->name('update')->middleware('admin');
+    Route::delete('/{media}', [PhotoController::class, 'destroy'])->name('destroy')->middleware('admin');
 });
 
 Route::group([
@@ -99,12 +100,12 @@ Route::group([
     'prefix' => 'videos'
 ], function () {
     Route::get('/', [VideoController::class, 'index'])->name('index');
-    Route::get('/create', [VideoController::class, 'create'])->name('create');
-    Route::post('/', [VideoController::class, 'store'])->name('store');
+    Route::get('/create', [VideoController::class, 'create'])->name('create')->middleware('admin');
+    Route::post('/', [VideoController::class, 'store'])->name('store')->middleware('admin');
     Route::get('/{media}/', [VideoController::class, 'show'])->name('show');
-    Route::get('/{media}/edit', [VideoController::class, 'edit'])->name('edit');
-    Route::put('/{media}', [VideoController::class, 'update'])->name('edit');
-    Route::delete('/{media}', [VideoController::class, 'destroy'])->name('destroy');
+    Route::get('/{media}/edit', [VideoController::class, 'edit'])->name('edit')->middleware('admin');
+    Route::put('/{media}', [VideoController::class, 'update'])->name('update')->middleware('admin');
+    Route::delete('/{media}', [VideoController::class, 'destroy'])->name('destroy')->middleware('admin');
 });
 
 Route::group([
@@ -112,12 +113,12 @@ Route::group([
     'prefix' => 'headings',
 ], function () {
     Route::get('/', [HeadingController::class, 'index'])->name('index');
-    Route::get('/create', [HeadingController::class, 'create'])->name('create');
-    Route::post('/', [HeadingController::class, 'store'])->name('store');
+    Route::get('/create', [HeadingController::class, 'create'])->name('create')->middleware('admin');
+    Route::post('/', [HeadingController::class, 'store'])->name('store')->middleware('admin');
     Route::get('/{heading}/', [HeadingController::class, 'show'])->name('show');
-    Route::get('/{heading}/edit', [HeadingController::class, 'edit'])->name('edit');
-    Route::put('/{heading}', [HeadingController::class, 'update'])->name('edit');
-    Route::delete('/{heading}', [HeadingController::class, 'destroy'])->name('destroy');
+    Route::get('/{heading}/edit', [HeadingController::class, 'edit'])->name('edit')->middleware('admin');
+    Route::put('/{heading}', [HeadingController::class, 'update'])->name('update')->middleware('admin');
+    Route::delete('/{heading}', [HeadingController::class, 'destroy'])->name('destroy')->middleware('admin');
 });
 
 Route::group([
@@ -125,32 +126,32 @@ Route::group([
    'prefix' => 'states'
 ], function () {
     Route::get('/', [StateController::class, 'index'])->name('index');
-    Route::get('/create', [StateController::class, 'create'])->name('create');
-    Route::post('/', [StateController::class, 'store'])->name('store');
+    Route::get('/create', [StateController::class, 'create'])->name('create')->middleware('admin');
+    Route::post('/', [StateController::class, 'store'])->name('store')->middleware('admin');
     Route::post('/search', [StateController::class, 'search'])->name('search');
     Route::get('/{state}/', [StateController::class, 'show'])->name('show');
-    Route::get('/{state}/edit', [StateController::class, 'edit'])->name('edit');
-    Route::put('/{state}', [StateController::class, 'update'])->name('update');
-    Route::delete('/{id}', [StateController::class, 'destroy'])->name('destroy');
-    Route::put('/recover/{id}', [StateController::class, 'recover'])->name('recover');
+    Route::get('/{state}/edit', [StateController::class, 'edit'])->name('edit')->middleware('admin');
+    Route::put('/{state}', [StateController::class, 'update'])->name('update')->middleware('admin');
+    Route::delete('/{id}', [StateController::class, 'destroy'])->name('destroy')->middleware('admin');
+    Route::put('/recover/{id}', [StateController::class, 'recover'])->name('recover')->middleware('admin');
 });
 
 Route::group([
     'as' => 'archives.',
     'prefix' => 'archives'
 ], function () {
-    Route::get('/', [ArchiveController::class, 'index'])->name('index');
-    Route::get('/create', [ArchiveController::class, 'create'])->name('create');
-    Route::post('/', [ArchiveController::class, 'store'])->name('store');
-    Route::post('/search', [ArchiveController::class, 'search'])->name('search');
-    Route::get('/{table}', [ArchiveController::class, 'show'])->name('show');
-    Route::get('/{table}/edit', [ArchiveController::class, 'edit'])->name('edit');
-    Route::put('/{table}', [ArchiveController::class, 'update'])->name('update');
-    Route::delete('/{table}/{id}', [ArchiveController::class, 'destroy'])->name('destroy');
-    Route::put('/{table}/{id}', [ArchiveController::class, 'recover'])->name('recover');
+    Route::get('/', [ArchiveController::class, 'index'])->name('index')->middleware('admin');
+    Route::get('/create', [ArchiveController::class, 'create'])->name('create')->middleware('admin');
+    Route::post('/', [ArchiveController::class, 'store'])->name('store')->middleware('admin');
+    Route::post('/search', [ArchiveController::class, 'search'])->name('search')->middleware('admin');
+    Route::get('/{table}', [ArchiveController::class, 'show'])->name('show')->middleware('admin');
+    Route::get('/{table}/edit', [ArchiveController::class, 'edit'])->name('edit')->middleware('admin');
+    Route::put('/{table}', [ArchiveController::class, 'update'])->name('update')->middleware('admin');
+    Route::delete('/{table}/{id}', [ArchiveController::class, 'destroy'])->name('destroy')->middleware('admin');
+    Route::put('/{table}/{id}', [ArchiveController::class, 'recover'])->name('recover')->middleware('admin');
 });
 
-Route::get('/addInfo', [AddInfoController::class, 'index'] )->name('addInfo');
+Route::get('/addInfo', [AddInfoController::class, 'index'] )->name('addInfo')->middleware('admin');
 
 Auth::routes();
